Cleanup, renaming and removing legacy code.
